对项目进行一个比较大的改动,主要是业务逻辑,编码问题等

+++ src/main/java/com/suny/association/mapper/interfaces/IMapper.java
@@ -16,11 +16,11 @@ public interface IMapper<T> {
 
     int insertAndGetId(T t);
 
-    void deleteById(int id);
+    void deleteById(long id);
 
     void update(T t);
 
-    T queryById(int id);
+    T queryById(long id);
 
     T queryByName(String name);
